a couple of warning fixes and support for assetcatalog:// as we migrate away from 320

+++ src/Three20UINavigator/Sources/TTURLNavigatorPattern.m
@@ -189,7 +189,7 @@ static NSString* kUniversalURLPattern = @"*";
 
     } else {
       char argType[256];
-      method_getArgumentType(method, wildcard.argIndex+2, argType, 256);
+      method_getArgumentType(method, (unsigned int)wildcard.argIndex+2, argType, 256);
       wildcard.argType = TTConvertArgumentType(argType[0]);
     }
   }
